feat(platform): extract the platform logic of the core (#30)
* feat(external): remove all spigot dependency to permits multiplatform

* feat(external): fix publishing

* feat(rework): rework architecture

* feat(version): fix versionning check

* feat(version): fix java version

* fix(version): jdk version of jitpack

* feat: change name of artifacts

* fix(typo): fix usage of newer java

* feat: remove Object usage to use generic type possibility
