+++ core/build.gradle
@@ -1,0 +1,43 @@
+plugins {
+    id 'maven-publish'
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+    withSourcesJar()
+    withJavadocJar()
+}
+
+def generatedResourcesDir = "$buildDir/generated-resources"
+
+task generateCommandsProperties {
+    doLast {
+        def outputDir = file("$generatedResourcesDir")
+        outputDir.mkdirs()
+        def propsFile = new File(outputDir, "commands.properties")
+        propsFile.text = "version=${project.version}"
+    }
+}
+
+processResources.dependsOn(generateCommandsProperties)
+
+// Inclure dans les ressources du build
+sourceSets {
+    main {
+        resources {
+            srcDir generatedResourcesDir
+        }
+    }
+}
+
+publishing {
+    publications {
+        maven(MavenPublication) {
+            from components.java
+            groupId = project.group
+            artifactId = 'core'
+            version = project.version
+        }
+    }
+}
