+++ core/src/main/java/fr/traqueur/commands/api/Arguments.java
