+++ core/src/main/java/fr/traqueur/commands/api/CommandPlatform.java
@@ -1,0 +1,18 @@
+package fr.traqueur.commands.api;
+
+import java.util.logging.Logger;
+
+public interface CommandPlatform<T, S> {
+
+    T getPlugin();
+
+    void injectManager(CommandManager<T, S> commandManager);
+
+    Logger getLogger();
+
+    boolean hasPermission(S sender, String permission);
+
+    void addCommand(Command<T, S> command, String label);
+
+    void removeCommand(String label, boolean subcommand);
+}
