+++ core/src/main/java/fr/traqueur/commands/api/arguments/Argument.java
