+++ core/src/main/java/fr/traqueur/commands/api/arguments/ArgumentValue.java
