+++ core/src/main/java/fr/traqueur/commands/api/exceptions/ArgsWithInfiniteArgumentException.java
