+++ core/src/main/java/fr/traqueur/commands/api/exceptions/ArgumentNotExistException.java
