+++ core/src/main/java/fr/traqueur/commands/api/logging/Logger.java
