+++ core/src/main/java/fr/traqueur/commands/api/logging/MessageHandler.java
