+++ core/src/main/java/fr/traqueur/commands/impl/arguments/DoubleArgument.java
