+++ core/src/main/java/fr/traqueur/commands/impl/arguments/LongArgument.java
