+++ core/src/main/java/fr/traqueur/commands/impl/logging/InternalLogger.java
