+++ core/src/main/java/fr/traqueur/commands/impl/logging/InternalMessageHandler.java
