+++ jitpack.yml
@@ -1,2 +1,6 @@
 jdk:
   - openjdk21
+
+build:
+  commands:
+    - ./gradlew core:publishToMavenLocal spigot:publishToMavenLocal
