+++ settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'CommandsAPI'
 
 include 'test-plugin'
+include 'spigot'
+include 'core'
