+++ spigot/src/main/java/fr/traqueur/commands/spigot/SpigotCommandContext.java
@@ -1,0 +1,11 @@
+package fr.traqueur.commands.spigot;
+
+import fr.traqueur.commands.api.Arguments;
+import fr.traqueur.commands.api.CommandContext;
+import org.bukkit.command.CommandSender;
+
+public class SpigotCommandContext extends CommandContext<CommandSender> {
+    public SpigotCommandContext(CommandSender sender, Arguments args) {
+        super(sender, args);
+    }
+}
