+++ spigot/src/main/java/fr/traqueur/commands/spigot/arguments/completions/TabCompleter.java
@@ -1,0 +1,6 @@
+package fr.traqueur.commands.spigot.arguments.completions;
+
+import org.bukkit.command.CommandSender;
+
+public interface TabCompleter extends fr.traqueur.commands.api.arguments.TabCompleter<CommandSender> {
+}
