--- src/main/resources/version.properties
@@ -1,1 +1,0 @@
-version=3.0.1
