+++ test-plugin/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'io.github.goooler.shadow' version '8.1.8'
+    id 'xyz.jpenilla.run-paper' version '2.3.1'
 }
 
 group = 'fr.traqueur'
@@ -18,9 +19,16 @@ repositories {
     }
 }
 
+tasks {
+    runServer {
+        minecraftVersion("1.21.3")
+    }
+}
+
 dependencies {
-    compileOnly("org.spigotmc:spigot-api:1.21.1-R0.1-SNAPSHOT")
-    implementation rootProject
+    compileOnly("org.spigotmc:spigot-api:1.21.3-R0.1-SNAPSHOT")
+    implementation project(":spigot")
+    implementation project(":core")
 }
 
 def targetJavaVersion = 21
