+++ test-plugin/src/main/java/fr/traqueur/testplugin/Sub2TestCommand.java
@@ -1,11 +1,9 @@
 package fr.traqueur.testplugin;
 
-import fr.traqueur.commands.api.Arguments;
-import fr.traqueur.commands.api.Command;
+import fr.traqueur.commands.api.CommandContext;
+import fr.traqueur.commands.spigot.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.plugin.java.JavaPlugin;
 
-import java.util.List;
 
 public class Sub2TestCommand extends Command<TestPlugin> {
 
@@ -15,9 +13,11 @@ public class Sub2TestCommand extends Command<TestPlugin> {
         this.addAlias("sub2.inner");
     }
 
+
     @Override
-    public void execute(CommandSender sender, Arguments args) {
-        args.getAsInt("test").ifPresent(test -> sender.sendMessage("Test: " + test));
-        sender.sendMessage(this.getUsage());
+    public void execute(CommandContext<CommandSender> context) {
+        context.args().getAsInt("test").ifPresent(test -> context.sender().sendMessage("Test: " + test));
+        context.sender().sendMessage(this.getUsage());
     }
+
 }
