+++ test-plugin/src/main/java/fr/traqueur/testplugin/SubTestCommand.java
@@ -1,9 +1,9 @@
 package fr.traqueur.testplugin;
 
 import fr.traqueur.commands.api.Arguments;
-import fr.traqueur.commands.api.Command;
+import fr.traqueur.commands.api.CommandContext;
+import fr.traqueur.commands.spigot.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.plugin.java.JavaPlugin;
 
 import java.util.List;
 
@@ -12,9 +12,9 @@ public class SubTestCommand extends Command<TestPlugin> {
     public SubTestCommand(TestPlugin plugin) {
         super(plugin, "sub.inner");
         this.addArgs("test");
-        this.addArgs("testStr", String.class, (sender, args) -> {
-           args.forEach(arg -> {
-               sender.sendMessage("Arg: " + arg);
+        this.addArgs("testStr", String.class, (context) -> {
+           context.args().forEach(arg -> {
+               context.sender().sendMessage("Arg: " + arg);
            });
            return List.of();
         });
@@ -22,10 +22,10 @@ public class SubTestCommand extends Command<TestPlugin> {
     }
 
     @Override
-    public void execute(CommandSender sender, Arguments args) {
-        int test = args.getAsInt("test", -1);
-        String testStr = args.get("testStr");
-        sender.sendMessage("Test: " + test + " TestStr: " + testStr);
-        sender.sendMessage(this.getUsage());
+    public void execute(CommandContext<CommandSender> context) {
+        int test = context.args().getAsInt("test", -1);
+        String testStr = context.args().get("testStr");
+        context.sender().sendMessage("Test: " + test + " TestStr: " + testStr);
+        context.sender().sendMessage(this.getUsage());
     }
 }
